Split service provider logic out into provider class
